Add `@lavamoat/allow-scripts`
Allow scripts is used to control which post-install scripts are run.
This protects against supply-chain attacks using post-install scripts.

A "setup" npm script was added to make it easier to install and run any
required post-install scripts with one script.

Contributor instructions have been added to the README to explain this
script, along with all other setup, testing, and linting instructions.

Also the `--frozen-lockfile` flag has been added to the install step in
CI as well. This should have been there already - it was accidentally
omitted.