Change `jsonSortOrder` option to a JSON string (#118)
The `jsonSortOrder` option has been converted from a path to a string.
The sort order object must now be specified as a JSON string rather
than in a separate file.

This should improve performance by avoiding additional interactions
with the filesystem. It also avoids having to resolve the JSON sort
order file path to the correct location, which is challenging if it's
specified as a CLI argument.