Migrate to Prettier v3
Prettier has been updated to v3. We now use types from Prettier
directly, so the types package has been dropped as well.

This plugin was affected by three changes mentioned in the changelog [1]:

* `parse` now has one less parameter. The `parsers` parameter (which
wasn't being used by this plugin) was removed.
* The Babel parser is now located at `prettier/plugins/babel`.
* Prettier now expects parsers to be async

A change was also required due to a seemingly-undocumented change to
the output of the Prettier JSON parser. The AST is now returned wrapped
in a `JsonRoot` node, which was not the case in v2.

`eslint-plugin-prettier` has been updated to v5, which is compatible
with Prettier v3.

fixes #119

[1]: https://prettier.io/blog/2023/07/05/3.0.0.html