Drop support for Node.js v16 (#203)
Support for Node.js v16 has been dropped. We no longer test v16 in CI,
and the `engines` field has been updated to set v18 as the minimum
supported version.

I've dropped support because Node.js v16 is no longer maintained, and
because this bump was necessary in order to use newer versions of Yarn
and AVA.

The README has been updated to suggest using the latest LTS for
development rather than the minimum version.