feat: Export `SortJsonOptions` and convert `CategorySort` to TS Union (#256)
The primary goal here is to export the `SortJsonOptions` interface so that I can strongly type my Prettier config including plugins.

In doing so, I figured we should probably export `CategorySort` as well. Instead, I converted the enum to a string union. This simplifies the syntax a bit, and enables TypeScript auto-complete on the strings without needing to import the enum. `CategorySort` has also been exported.