Improve documentation around `json-sort-order` option (#279)
The `json-sort-order` documentation has been reorganized and expanded.
The example given has been simplified and explained in much more
detail.

Hopefully it will now be more clear to users that a regular expression
rule must start with `/`. The simplest regular expression rule possible
(`/.*/`) is now used directly in the example as well.

Relates to #275