Add recommendation to avoid escaping (#280)
Update the "Escaping" section of the README to better describe why
escaping is needed, what hazards to look out for, and how to avoid
them.

Previously the reasoning for the format was not explained, which led to
some confusion about this design choice (why would we make this so
difficult?). A small amount of context has been added to help explain
this.

Also, previously it wrongly suggested that double-escaping was only
needed for JSON configuration. It's actually required for configuration
in JavaScript files as well, if the value is specified as a string.
The double-escaping is required if the option is represented as a
string in any format.

Lastly, a recommendation was added to allow users to avoid all escaping
in their configuration files.